Fix -Wc++11-narrowing: rlz/
Bug: 1216696
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2975075
Auto-Submit: Peter Kasting <pkasting@chromium.org>
Commit-Queue: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/master@{#894214}

+++ rlz/lib/rlz_lib.cc
@@ -445,10 +445,12 @@ bool SendFinancialPing(Product product,
 #if BUILDFLAG(IS_CHROMEOS_ASH)
   const net::BackoffEntry::Policy policy = {
       0,  // Number of initial errors to ignore.
-      base::TimeDelta::FromSeconds(5).InMilliseconds(),  // Initial delay.
+      static_cast<int>(
+          base::TimeDelta::FromSeconds(5).InMilliseconds()),  // Initial delay.
       2,    // Factor to increase delay.
       0.1,  // Delay fuzzing.
-      base::TimeDelta::FromMinutes(5).InMilliseconds(),  // Maximum delay.
+      static_cast<int>(
+          base::TimeDelta::FromMinutes(5).InMilliseconds()),  // Maximum delay.
       -1,  // Time to keep entries.  -1 == never discard.
   };
   net::BackoffEntry backoff(&policy);
